Create SelectionActionMixin and refactor d2l-selection-action and d2l-selection-action-menu-item to use it.
