+++ components/selection/selection-action-menu-item.js
@@ -1,8 +1,8 @@
 import { css, html, LitElement } from 'lit-element/lit-element.js';
 import { MenuItemMixin } from '../menu/menu-item-mixin.js';
 import { menuItemStyles } from '../menu/menu-item-styles.js';
+import { SelectionActionMixin } from './selection-action-mixin.js';
 import { SelectionInfo } from './selection-mixin.js';
-import { SelectionObserverMixin } from './selection-observer-mixin.js';
 
 /**
  * An action menu-item component used within selection controls such as d2l-list and d2l-list-header.
@@ -10,17 +10,7 @@ import { SelectionObserverMixin } from './selection-observer-mixin.js';
  * @fires d2l-selection-action-click - Dispatched when the user clicks the action button. The `SelectionInfo` is provided as the event `detail`. If `requires-selection` was specified then the event will only be dispatched if items are selected.
  * @fires d2l-selection-observer-subscribe - Internal event
  */
-class MenuItem extends SelectionObserverMixin(MenuItemMixin(LitElement)) {
-
-	static get properties() {
-		return {
-			/**
-			 * Whether the action menu-item requires one or more selected items
-			 * @type {boolean}
-			 */
-			requiresSelection: { type: Boolean, attribute: 'requires-selection', reflect: true }
-		};
-	}
+class MenuItem extends SelectionActionMixin(MenuItemMixin(LitElement)) {
 
 	static get styles() {
 		return [ menuItemStyles,
@@ -36,15 +26,6 @@ class MenuItem extends SelectionObserverMixin(MenuItemMixin(LitElement)) {
 				}
 			`
 		];
-	}
-
-	get selectionInfo() {
-		return super.selectionInfo;
-	}
-
-	set selectionInfo(value) {
-		super.selectionInfo = value;
-		this.disabled = (this.requiresSelection && this.selectionInfo.state === SelectionInfo.states.none);
 	}
 
 	connectedCallback() {
